support_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
PyJWT
email-validator
pydantic[email]
pydantic-settings